Move binary resolution to separate module
