+++ lib/client/index.ts
@@ -1,4 +1,5 @@
-export {ClientResolver, resolveBinary} from "./clientResolver"
+export {ClientResolver} from "./clientResolver"
+export {resolveBinary} from "./resolveBinary"
 
 import {TypescriptServiceClient} from "./client"
